fix(sdk-ui-filters): irrelevant values when apply all at once
- in attribute filters

JIRA: MC-3609
risk: low
