+++ common/changes/@gooddata/sdk-ui-all/ovel-fix-dependent-filter-in-apply-all_2025-05-13-13-21.json
@@ -1,0 +1,10 @@
+{
+    "changes": [
+        {
+            "packageName": "@gooddata/sdk-ui-all",
+            "comment": "Fix irrelevant values info in attribute filter when apply all at once is on",
+            "type": "none"
+        }
+    ],
+    "packageName": "@gooddata/sdk-ui-all"
+}
