+++ libs/sdk-ui-filters/src/AttributeFilterHandler/internal/redux/common/selectors.ts
@@ -1,4 +1,4 @@
-// (C) 2021-2024 GoodData Corporation
+// (C) 2021-2025 GoodData Corporation
 import { IAttributeElement } from "@gooddata/sdk-model";
 import { AttributeFilterState } from "../store/state.js";
 
@@ -23,3 +23,8 @@ export const getElementCacheKey = (
  * @internal
  */
 export const selectElementsForm = (state: AttributeFilterState) => state.elementsForm;
+
+/**
+ * @internal
+ */
+export const selectWithoutApply = (state: AttributeFilterState) => state.config.withoutApply;
